+++ task-management/security/cors/middlewares/corsMiddleware.js
@@ -1,5 +1,5 @@
 import cors from 'cors'
-import corsOptions from './corsOptions.js'
+import corsOptions from '../config/corsOptions.js'
 
 const corsMiddleware = cors(corsOptions) // Middleware para manejar CORS
